added Daily report component
Daily component now returns a 7 days forecast with basic information.  To do so it needs latitude and longitute of the citiy to return the values. The background color for the daily cards would be also determine by the weather condition.

added a export to the iconKey function so it can be used by other componets.
removed the card's background gradient color by a solid color background instead.