Daily component UI
added transitions to the daily report component when hiding and displaying the panels, the "more" button has been removed by an arrow icon and moved the click function to the parent element.

Added react icons into the project.

changed the way that the app fetches the data from the APIs. the loading variable will toggle to false when all the data has been fetched from the APIs.

created css folder to store stylesheets in the future. removed the css that affects the daily component into a new file for better reading.